renames property in cache trait comment; casts the cache count to int

+++ src/app/Classes/Table/Builder.php
@@ -327,7 +327,7 @@ class Builder
             );
         }
 
-        return cache()->get(
+        return (int) cache()->get(
             'datatable:'.$this->request->get('name')
         );
     }
